+++ packages/expo-calendar/android/src/main/java/expo/modules/calendar/domain/event/records/Alarm.kt
@@ -1,0 +1,10 @@
+package expo.modules.calendar.domain.event.records
+
+import expo.modules.calendar.domain.event.enums.AlarmMethod
+import expo.modules.kotlin.records.Field
+import expo.modules.kotlin.records.Record
+
+data class Alarm(
+  @Field val relativeOffset: Int?,
+  @Field val method: AlarmMethod?
+) : Record